+++ src/backend/src/transformers/proposed-solutions.transformer.ts
@@ -1,7 +1,7 @@
 import { Prisma } from '@prisma/client';
 import { ProposedSolution } from 'shared';
 import proposedSolutionArgs from '../prisma-query-args/proposed-solutions.query-args';
-import userTransformer from './user.transformer';
+import { userTransformer } from './user.transformer';
 
 const proposedSolutionTransformer = (
   proposedSolution: Prisma.Proposed_SolutionGetPayload<typeof proposedSolutionArgs>
