+++ src/backend/src/transformers/user.transformer.ts
@@ -1,7 +1,8 @@
 import { Prisma } from '@prisma/client';
-import { User } from 'shared';
+import { User, UserWithScheduleSettings } from 'shared';
+import userScheduleSettingsTransformer from './user-schedule-settings.transformer';
 
-const userTransformer = (user: Prisma.UserGetPayload<null>): User => {
+export const userTransformer = (user: Prisma.UserGetPayload<null>): User => {
   return {
     userId: user.userId,
     firstName: user.firstName,
@@ -12,4 +13,16 @@ const userTransformer = (user: Prisma.UserGetPayload<null>): User => {
   };
 };
 
-export default userTransformer;
+export const userWithScheduleSettingsTransformer = (
+  user: Prisma.UserGetPayload<{ include: { drScheduleSettings: true } }>
+): UserWithScheduleSettings => {
+  return {
+    userId: user.userId,
+    firstName: user.firstName,
+    lastName: user.lastName,
+    email: user.email,
+    emailId: user.emailId,
+    role: user.role,
+    scheduleSettings: user.drScheduleSettings ? userScheduleSettingsTransformer(user.drScheduleSettings) : undefined
+  };
+};
