+++ src/backend/tests/test-data/reimbursement-requests.test-data.ts
@@ -15,8 +15,8 @@ import { alfred, batman } from './users.test-data';
 import { prismaWbsElement1 } from './wbs-element.test-data';
 import { ClubAccount, ReimbursementRequest } from 'shared';
 import { wbsNumOf } from '../../src/utils/utils';
-import userTransformer from '../../src/transformers/user.transformer';
 import { vendorTransformer, expenseTypeTransformer } from '../../src/transformers/reimbursement-requests.transformer';
+import { userTransformer } from '../../src/transformers/user.transformer';
 
 export const PopEyes: PrismaVendor = {
   vendorId: 'CHICKEN',
